#93: [Refator] 時間管理ロジックをカスタムフックに分離
